Modularize project structure
